chore: combine handler and main
